requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
transformers==4.37.2
accelerate==0.23.0
bitsandbytes==0.41.1
datasets==2.14.5
peft==0.5.0
scikit-learn==1.3.1
sentencepiece==0.1.99
trl==0.7.2
protobuf==3.20.3
ipykernel==6.25.2
wandb==0.15.12
onnxruntime-gpu==1.17.1
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-genai/pypi/simple/
onnxruntime-genai-cuda
einops
promptflow==0.1.0b8
promptflow-tools==0.1.0b10
git+https://github.com/microsoft/Olive